Reorganize SoS container modules.
